backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
python-dotenv==1.2.1
mysql-connector-python==9.5.0
joblib==1.3.2
numpy==1.26.2
scikit-learn==1.6.1